+++ SunsynkIHD/lib/Displays/Displays.h
@@ -1,0 +1,10 @@
+#ifndef __DISPLAYS_H
+#define __DISPLAYS_H
+
+#include "esp32_ili9488_spi.h"
+#include "esp32s3_elecrow_hmi_35.h"
+#include "esp32s3_elecrow_hmi_70.h"
+
+extern LGFX *gfx;
+
+#endif // __DISPLAYS_H
