--- SunsynkIHD/lib/FT6236/library.properties
@@ -1,10 +1,0 @@
-name=FT6236 Library
-version=1.0.2
-author=Dustin Watts
-maintainer=Dustin Watts<info@dustinwatts.nl>
-sentence=Arduino library for the FocalTech FT6236 touch controller
-paragraph=Arduino library for the FocalTech FT6236 touch controller
-category=Display
-url=https://github.com/DustinWatts/FT6236
-architectures=*
-depends=Wire
