+++ platformio.ini
@@ -18,10 +18,8 @@ platform = espressif32
 board = esp32dev
 framework = arduino
 lib_deps = 
-	moononournation/GFX Library for Arduino@^1.3.6
 	lvgl/lvgl@^8.3.7
 	bblanchon/ArduinoJson@^6.21.2
-	paulstoffregen/XPT2046_Touchscreen@0.0.0-alpha+sha.26b691b2c8
 	paulstoffregen/Time@^1.6.1
 	lovyan03/LovyanGFX@^1.1.7
 build_flags = 
@@ -38,7 +36,6 @@ platform = espressif32
 board = elecrow-esp32s3-hmi-35
 framework = arduino
 lib_deps = 
-	moononournation/GFX Library for Arduino@^1.3.6
 	lvgl/lvgl@^8.3.7
 	bblanchon/ArduinoJson@^6.21.2
 	paulstoffregen/Time@^1.6.1
@@ -56,11 +53,9 @@ board = elecrow-esp32s3-hmi-70
 board_build.partitions = huge_app.csv
 framework = arduino
 lib_deps = 
-	moononournation/GFX Library for Arduino@^1.3.6
 	lvgl/lvgl@^8.3.7
 	bblanchon/ArduinoJson@^6.21.2
 	paulstoffregen/Time@^1.6.1
-	tamctec/TAMC_GT911@^1.0.2
 	lovyan03/LovyanGFX@^1.1.7
 build_flags = 
 	-DCORE_DEBUG_LEVEL=2
